# api.yml
openapi: 3.0.0
info:
  version: 1.0.0
  title: YardRaffles
servers:
  - url: 'http://localhost:8888'
paths:
...
  /api/admin/getAdminUser:
    post:
      tags:
        - admin
      summary: "Получить информацию о пользователе"
      operationId: getAdminUser
      parameters:
      - name: authorization
        in: header
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
            format: hash
      - name: email
        in: query
        required: true
        schema:
            type: integer
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/UserData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Пользователь не найден в БД"
  # AUTH END
  
  # BOT START
  /api/bot/getUser/{telegramId}:
    get:
      tags:
        - bot
      summary: "Получить информацию о пользователе"
      operationId: getBotUser
      parameters:
      - name: Authorization
        in: header
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - name: telegramId
        in: path
        required: true
        description: telegram_id пользователя
        schema:
            type: integer
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/UserData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Пользователь не найден в БД"
  /api/bot/addUser:
    post:
      tags:
        - bot
      summary: "Добавляет нового пользователя в боте"
      operationId: addBotUser
      parameters:
      - name: Authorization
        in: header
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - name: username
        in: query
        required: true
        schema:
            type: string
      - name: telegram_link
        in: query
        required: true
        schema:
            type: string
      - in: query
        name: phone
        required: false
        schema:
            type: string
      - in: query
        name: email
        schema:
            type: string
        required: false
      - in: query
        name: telegramId
        schema:
            type: integer
        required: true
        description: telegram_id пользователя
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/UserData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Пользователь не найден в БД"
  /api/bot/updateUser/{telegramId}:
    put:
      tags:
        - bot
      summary: "Обновляет информацию пользователя в боте"
      operationId: updateBotUser
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: query
        name: username
        schema:
            type: string
        required: true
      - in: query
        name: telegram_link
        schema:
            type: string
        required: true
      - in: query
        name: phone
        schema:
            type: string
        required: false
      - in: query
        name: email
        schema:
            type: string
        required: false
      - in: path
        name: telegramId
        schema:
            type: integer
        required: true
        description: telegram_id пользователя
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/UserData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Пользователь не найден в БД"
  /api/bot/getRafflesToPost:
    get:
      tags:
        - bot
      summary: "Получает список раффлов для публикации"
      operationId: getBotRafflesToPost
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
                schema:
                    type: object
                    properties:
                        data:
                            type: array
                            items:
                                $ref: '#/definitions/BotRaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Раффл не найден в БД"
  /api/bot/getCurrentRaffles:
    get:
      tags:
        - bot
      summary: "Получает список текущих раффлов"
      operationId: getBotCurrentRaffles
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                type: object
                properties:
                    data:
                        type: array
                        items:
                            $ref: '#/definitions/BotRaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Раффл не найден в БД"
  /api/bot/getRaffle/{id}:
    get:
      tags:
        - bot
      summary: "Получает раффл"
      operationId: getRaffle
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: id раффла
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/BotRaffleData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
        404:
          description: "Раффл не найден в БД"
  /api/bot/addPartRaffle/{id}:
    post:
      tags:
        - bot
      summary: "Регестрирует на раффл"
      operationId: addPartRaffle
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: query
        name: telegramId
        schema:
            type: integer
        required: true
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: id раффла
      responses:
        200:
          description: "Успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/bot/confirmParticipation/{id}:
    put:
      tags:
        - bot
      summary: "Обновляет статус записи"
      operationId: confirmParticipation
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: query
        name: telegramId
        schema:
            type: integer
        required: true
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: id раффла
      responses:
        200:
          description: "Успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/bot/confirmSuccess/{id}:
    put:
      tags:
        - bot
      summary: "Обновляет статус записи"
      operationId: confirmSuccess
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) д. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        schema:
            type: string
      - in: query
        name: telegramId
        schema:
            type: integer
        required: true
      - in: path
        name: id
        schema:
            type: integer
        required: true
        description: id раффла
      responses:
        200:
          description: "Объект с записью"
          content:
            application/json:
              schema:
                $ref: '#/definitions/SuccessData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  # BOT END
  
  # USERS START
...
  /api/announces/create:
    post:
      tags:
      - announcements
      summary: Создать объялвение
      operationId: createAnnounce
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: query
        name: title
        type: string
        required: true
        description: "Название"
      - in: query
        name: work_name
        type: string
        required: true
        description: "Рабочее название"
      - in: query
        name: images
        type: string
        required: false
        description: "Hash изображения"
      - in: query
        name: message
        type: string
        required: true
        description: "Сообщение в боте"
      - in: query
        name: publication_date
        type: string
        required: false
        description: "Время публикации (если не указано, то сразу)"
      - in: body
        name: tags
        required: false
        description: "Массив объектов id тэгов"
        schema:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
      responses:
        200:
          description: "Объект с созданной записью"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/AnnounceData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/announces/update/{id}:
    put:
      tags:
      - announcements
      summary: Обновить объялвение
      operationId: updateAnnounce
      produces:
      - application/json
      parameters:
      - in: header
        name: Authorization
        description: "hash пользователя (токен авторизации) для расчета стоимости индивидуального посещения. Пример: Bearer c4ca4238a0b923820dcc509a6f75849b"
        required: true
        type: string
      - in: path
        name: id
        type: integer
        required: true
        description: Id раффла
      - in: query
        name: title
        type: string
        required: true
        description: "Название"
      - in: query
        name: work_name
        type: string
        required: true
        description: "Рабочее название"
      - in: query
        name: images
        type: string
        required: false
        description: "Hash изображения"
      - in: query
        name: message
        type: string
        required: true
        description: "Сообщение в боте"
      - in: query
        name: publication_date
        type: string
        required: false
        description: "Время публикации (если не указано, то сразу)"
      - in: body
        name: tags
        required: false
        description: "Массив объектов id тэгов"
        schema:
          type: array
          items:
            type: integer
          example: [1, 2, 3]
      responses:
        200:
          description: "Объект с созданной записью"
          schema:
            type: object
            properties:
              data:
                $ref: '#/definitions/AnnounceData'
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
  /api/announces/delete/{id}:
    delete:
      tags:
      - announcements
      summary: Удалить запись объялвения
      operationId: deleteAnnounce
      produces:
      - application/json
      parameters:
      - in: path
        name: id
        type: integer
        required: true
        description: Id записи
        
      responses:
        200:
          description: "Операция прошла успешно"
        401:
          description: "Токен не отправлен или пользователь с таким токеном не найден в БД"
...